move code from main
